5/20 Progress
Includes my object oriented approach to  programming the blaster. Ultimately, I had to abandon this approach because of weird characters appearing on the LCD screen when being implemented with an LCD class. I tried turning off interrupts. This is a problem I will solve in the future. The functional approach I am working on now is more efficient and intuitive for Arduino. My tests for the classes are also included. Also a test for the PIR sensor located in the barrel.